requirements.txt
Flask==2.3.2
openai==0.28.0
python-dotenv==1.0.0
gunicorn
